# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniSight RSM",
    "description": "Ресурсно-сервисная модель (РСМ) — API v1 для учёта элементов, связей и корреляции алертов.\n\n**Integration API (ТЗ §8)** — `Resources`, `Correlation`, `Meta`: контракт для correlation engine и внешних систем мониторинга.\n\n**Admin API** — `CI`, `Relations`, справочники, Autodiscover, аудит: ведение модели через UI и администрирование.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/correlation/chain-check": {
      "post": {
        "tags": [
          "Integration API · Correlation (§8)"
        ],
        "summary": "Correlation Chain Check V1",
        "operationId": "correlation_chain_check_v1_api_v1_correlation_chain_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChainCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/correlation/context": {
      "post": {
        "tags": [
          "Integration API · Correlation (§8)"
        ],
        "summary": "Correlation Context V1",
        "operationId": "correlation_context_v1_api_v1_correlation_context_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrelationContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/correlation/ingest": {
      "post": {
        "tags": [
          "Integration API · Correlation (§8)"
        ],
        "summary": "Correlation Ingest V1",
        "operationId": "correlation_ingest_v1_api_v1_correlation_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrelationIngestRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/ci/types": {
      "get": {
        "tags": [
          "Admin API · CI Types"
        ],
        "summary": "List Ci Types V1",
        "operationId": "list_ci_types_v1_api_v1_ci_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin API · CI Types"
        ],
        "summary": "Create Ci Type V1",
        "operationId": "create_ci_type_v1_api_v1_ci_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CITypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/ci/types/{type_id}": {
      "patch": {
        "tags": [
          "Admin API · CI Types"
        ],
        "summary": "Patch Ci Type V1",
        "operationId": "patch_ci_type_v1_api_v1_ci_types__type_id__patch",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CITypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin API · CI Types"
        ],
        "summary": "Remove Ci Type V1",
        "operationId": "remove_ci_type_v1_api_v1_ci_types__type_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/relation/types": {
      "get": {
        "tags": [
          "Admin API · Relation Types"
        ],
        "summary": "List Relation Types V1",
        "operationId": "list_relation_types_v1_api_v1_relation_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin API · Relation Types"
        ],
        "summary": "Create Relation Type V1",
        "operationId": "create_relation_type_v1_api_v1_relation_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationTypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/relation/types/{type_id}": {
      "patch": {
        "tags": [
          "Admin API · Relation Types"
        ],
        "summary": "Patch Relation Type V1",
        "operationId": "patch_relation_type_v1_api_v1_relation_types__type_id__patch",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationTypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin API · Relation Types"
        ],
        "summary": "Remove Relation Type V1",
        "operationId": "remove_relation_type_v1_api_v1_relation_types__type_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertIdentifier": {
        "properties": {
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "serviceCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servicecode"
          },
          "applicationCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicationcode"
          },
          "externalId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Externalid"
          },
          "cmdbId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmdbid"
          }
        },
        "type": "object",
        "title": "AlertIdentifier"
      },
...
      "CITypeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "attribute_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribute Schema"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CITypeCreate"
      },
      "CITypeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "attribute_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribute Schema"
          }
        },
        "type": "object",
        "title": "CITypeUpdate"
      },
...
      "ChainCheckRequest": {
        "properties": {
          "resource_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Resource Ids"
          }
        },
        "type": "object",
        "required": [
          "resource_ids"
        ],
        "title": "ChainCheckRequest",
        "description": "FR 39: проверка принадлежности объектов одной depends_on-цепочке."
      },
      "ChainCheckResponse": {
        "properties": {
          "resource_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Resource Ids"
          },
          "chain_related": {
            "type": "boolean",
            "title": "Chain Related"
          },
          "chain_algorithm": {
            "type": "string",
            "title": "Chain Algorithm",
            "default": "depends_on_directed"
          }
        },
        "type": "object",
        "required": [
          "resource_ids",
          "chain_related"
        ],
        "title": "ChainCheckResponse"
      },
...
      "CorrelationContextRequest": {
        "properties": {
          "resource_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Resource Ids"
          },
          "depth": {
            "type": "integer",
            "title": "Depth",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "resource_ids"
        ],
        "title": "CorrelationContextRequest"
      },
      "CorrelationIngestRequestV1": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 100
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertIdentifier"
            },
            "type": "array",
            "title": "Alerts"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "depth": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Depth",
            "default": 3
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "dispatch_webhook": {
            "type": "boolean",
            "title": "Dispatch Webhook",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "alerts"
        ],
        "title": "CorrelationIngestRequestV1",
        "examples": [
          {
            "alerts": [
              {
                "hostname": "app-01"
              },
              {
                "ip": "10.0.0.5"
              }
            ],
            "depth": 3,
            "dispatch_webhook": true,
            "page": 1,
            "page_size": 100,
            "source": "monitoring-system",
            "webhook_url": "https://monitoring.example/hooks/rsm"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RelationTypeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RelationTypeCreate"
      },
      "RelationTypeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "RelationTypeUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT из `POST /api/v1/auth/token` (поле `session.access_token`). В Authorize вставьте только токен."
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "Секрет webhook-интеграций."
      },
      "XApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Ключ интеграции для correlation ingest и внешних систем."
      }
    }
  },
...
}